Change root of app
