+++ src/i18n/errorMessage.js
@@ -3,6 +3,10 @@ const errorMessage = {
   questionnaireNotFound: { fr: 'Questionnaire introuvable', en: 'Questionnaire not found' },
   surveyUnitNotFound: { fr: 'Unité enquêtée introuvable', en: 'Survey unit not found' },
   putSurveyUnitFailed: { fr: "Échec lors de l'envoi des données", en: 'Failed to send data' },
+  unauthorized: {
+    fr: "Vous n'êtes pas autorisé à accéder à l'application",
+    en: 'You are not authorized to access the application.',
+  },
 };
 
 export default errorMessage;
